feat: apply code quality conventions and fix critical violations (#29)
- Add explicit return types to all functions in App.tsx
- Fix import order to follow conventions (external → relative → types)
- Extract CutscenePlayer types to separate types.ts file
- Move mock data to dedicated mocks.ts file
- Remove console.error statements from production code
- Update conventions.md with identified gaps:
  - Error handling strategy
  - Mock data organization
  - Logging strategy
  - Async operation patterns
  - Performance documentation guidelines

Created follow-up issues for non-critical violations:
- #30: Move UI state to Zustand store
- #31: Improve documentation standards
- #32: Implement consistent error handling patterns
- #33: Standardize TypeScript interface vs type usage
- #34: Update test patterns to follow conventions
