+++ src/components/CutscenePlayer/index.ts
@@ -1,4 +1,4 @@
 export { CutscenePlayer, CutscenePlayerWithErrorBoundary } from './CutscenePlayer';
-export type { CutscenePlayerProps, CutsceneDefinition, CutsceneShot, PlayerState } from './CutscenePlayer';
 export { CutsceneErrorBoundary } from './components/CutsceneErrorBoundary';
 export { useCutscenePlayer } from './hooks/useCutscenePlayer';
+export type { CutscenePlayerProps, CutsceneDefinition, CutsceneShot, PlayerState } from './types';
